it goes boys... terribly, it it goes

+++ dataset/sudoku.py
@@ -115,8 +115,8 @@ class SudokuDataset(torch.utils.data.Dataset):
 
     def __getitem__(self, idx):
         return {
-            'puzzle': torch.tensor(self.puzzles[idx], dtype=torch.long),
-            'solution': torch.tensor(self.solutions[idx], dtype=torch.long)
+            'puzzle': torch.tensor(self.puzzles[idx], dtype=torch.float32),
+            'solution': torch.tensor(self.solutions[idx], dtype=torch.float32)
         }
